apps/server/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "video_audio_server" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/apps/server/html'
 --cov-report xml:'../../coverage/apps/server/coverage.xml'
 --html='../../reports/apps/server/unittests/html/index.html'
 --junitxml='../../reports/apps/server/unittests/junit.xml'
"""
asyncio_mode = "auto"

[tool.poetry]
name = "server"
version = "1.0.0"
description = "Video upload and audio transformation service"
authors = []
license = "Proprietary"
readme = "README.md"

  [[tool.poetry.packages]]
  include = "video_audio_server"

  [tool.poetry.dependencies]
  python = ">=3.12,<4.0"
  fastapi = { version = "^0.115.0", extras = ["standard"] }
  sqlalchemy = { version = "^2.0.0", extras = ["asyncio"] }
  greenlet = "^3.0.0"
  alembic = "^1.13.0"
  asyncpg = "^0.30.0"
  arq = "^0.26.0"
  redis = { version = "^5.0.0", extras = ["hiredis"] }
  aioboto3 = "^13.0.0"
  pyjwt = "^2.8.0"
  passlib = { version = "^1.7.4", extras = ["bcrypt"] }
  pydantic-settings = "^2.3.0"
  numpy = "^2.0.0"
  httpx = "^0.27.0"
  python-multipart = "^0.0.9"
  uvicorn = { version = "^0.30.0", extras = ["standard"] }

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.3.1"
  ruff = "0.8.2"
  pytest = "8.3.4"
  pytest-sugar = "1.0.0"
  pytest-cov = "6.0.0"
  pytest-html = "4.1.1"
  pytest-asyncio = "^0.24.0"
  mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "dist",
]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
  "E",
  "F",
  "UP",
  "B",
  "SIM",
  "I",
]
ignore = [
  "B008", # FastAPI uses Depends() in function signatures — this is idiomatic DI
]
fixable = ["ALL"]
unfixable = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
explicit_package_bases = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
warn_required_dynamic_aliases = true

[tool.pyright]
pythonVersion = "3.12"
reportUntypedBaseClass = "none"
